backend/app/requirements.txt
alembic==1.11.1
asyncpg==0.28.0
fastapi==0.100.0
fastapi-pagination==0.12.5
passlib==1.7.4
python-dotenv==1.0.0
python-jose==3.3.0
python-multipart
pytest==7.4.0
pytest-asyncio==0.21.0
pydantic==1.10.11
pydantic[email]==1.10.11
PyJWT==2.7.0
uvicorn==0.22.0
sqlalchemy==2.0.18
httpx==0.24.1
requests==2.31.0
types-requests==2.31.0.1


# LINTING
black==23.3.0
isort==5.12.0
pylint==2.13.9
mypy==1.3.0

# REQUIRED FOR LINTING SQLALCHEMY
# IGNORE THE FLASK - IT IS COMPATIBLE WITH FASTAPI.

pylint-flask-sqlalchemy
sqlalchemy[mypy]
types-python-jose
types-passlib

# dev tooling
faker==15.3.4
coverage

# celery
celery==5.2.7
psycopg==3.1.8
psycopg-binary==3.1.8
redis==4.5.5

